Removed async issues with watch task
The dependencies in a gulp task are not run synchronously or
necessarily in order, so using the functionality for restarting the
server does not work.

All tasks have been updated to use one of the accepted methods for
synchronisation, and the server restart has been moved to the callback
for the tasks.