Issue #130 - Docker compose (#141)
* Issue #130 - Docker compose

Migrated from fig to docker-compose. The dockerfile has been updated
to use the 0.10 image rather than onbuild to hopefully make it a little
easier to try some of the newer node images.

* Moved application to /opt directory

Updated the dockerfile to use the working directory, rather than
stating it repeatedly.

* Moved daemon functions into gulpfile

Instead of using nodemon, the server is started and restarted within
gulp. This helps with the issue of nodemon picking up changes while
tasks are in flight.

Also added the shrinkwrap file into the image to ensure we get the
correct package versions.

* Config added to docker ignore exceptions

The image builds a dev environment, so should contain the config
necessary to lint the input.

* Removed async issues with watch task

The dependencies in a gulp task are not run synchronously or
necessarily in order, so using the functionality for restarting the
server does not work.

All tasks have been updated to use one of the accepted methods for
synchronisation, and the server restart has been moved to the callback
for the tasks.